requirements.txt

Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
marshmallow==3.20.1
bleach==6.1.0
